fastapi-cattle-detection/requirements.txt
fastapi
uvicorn
torch
torchvision
pydantic
numpy
pandas
scikit-learn
Pillow
joblib
tqdm
python-dotenv
opencv-python
fastapi-utils